+++ examples/example_test.go
@@ -2,12 +2,15 @@ package main
 
 import (
 	"testing"
+	"time"
 
 	"gitlab.paradise-soft.com.tw/dwh/dispatcher/glob"
 	"gitlab.paradise-soft.com.tw/dwh/dispatcher/service"
 )
 
 func TestIntegration(t *testing.T) {
+	service.TopicService.Remove(glob.Config.Topic)
+	time.Sleep(time.Second)
 	type args struct {
 		msgCount int
 	}
@@ -17,6 +20,7 @@ func TestIntegration(t *testing.T) {
 		want int
 	}{
 		{"10 Messages", args{10}, 10},
+		{"15 Messages", args{15}, 15},
 		{"15 Messages", args{15}, 15},
 		// {"50k Messages", args{50000}, 50000},
 		// {"500k Messages", args{500000}, 500000},
@@ -31,6 +35,9 @@ func TestIntegration(t *testing.T) {
 		})
 	}
 	service.TopicService.Remove(glob.Config.Topic)
+	// service.TopicService.Remove(glob.Config.Topic+"_ERR")
+	// service.TopicService.Remove("disp.test.1")
+	// service.TopicService.Remove("disp.test.2")
 }
 
 func BenchmarkProducer(b *testing.B) {
@@ -38,4 +45,24 @@ func BenchmarkProducer(b *testing.B) {
 		Producer(glob.Config.Topic, 1)
 	}
 	service.TopicService.Remove(glob.Config.Topic)
+}
+
+func TestMultiConsProds(t *testing.T) {
+	type args struct {
+		msgCount int
+	}
+	tests := []struct {
+		name string
+		args args
+		want int
+	}{
+		{"10 Messages", args{10}, 20},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if got := MultiConsProds(tt.args.msgCount); got != tt.want {
+				t.Errorf("MultiConsProds() = %v, want %v", got, tt.want)
+			}
+		})
+	}
 }