+++ glob/init.go
@@ -66,12 +66,12 @@ func (c *config) loadSaramaConfigs() {
 	// Producer
 	tmpC.Producer.RequiredAcks = sarama.WaitForAll // Wait for all in-sync replicas to ack the message
 	tmpC.Producer.Retry.Max = 10                   // Retry up to 10 times to produce the message
-	// tmpC.Producer.Return.Successes = true       // Receive success msg
+	tmpC.Producer.Return.Successes = true          // Receive success msg
 
 	// Consumer
 	tmpC.Consumer.Return.Errors = true
 	tmpC.Consumer.Offsets.Initial = sarama.OffsetNewest
-	// tmpC.Consumer.Offsets.Initial = sarama.
+	// tmpC.Consumer.Offsets.Initial = sarama.OffsetOldest
 
 	// TLS
 	tlsConfig := c.createTlsConfiguration()