+++ model/consumer.go
@@ -8,6 +8,5 @@ type ConsumerCallback func(key, value []byte) error
 
 type ConsumerCallbackError struct {
 	Message *sarama.ConsumerMessage
-	Err     error
-	// Topic   string
+	ErrStr  string
 }