+++ service/producer_service.go
@@ -1,6 +1,9 @@
 package service
 
 import (
+	"encoding/json"
+	"errors"
+	"runtime/debug"
 	"sync"
 
 	"gitlab.paradise-soft.com.tw/dwh/dispatcher/model"
@@ -18,15 +21,15 @@ type producerService struct {
 	lock     *sync.Mutex
 }
 
-func (p *producerService) Send(topic string, key, value []byte, errHandler model.ProducerErrHandler) {
+func (p *producerService) Send(topic string, key, value []byte, customErrHandler model.ProducerCustomerErrHandler) {
 	p.send(topic, key, value)
-	// ConsummerService.Subscribe(topic, callback, asyncNum)
+	ConsumerService.Subscribe(glob.ErrTopic(topic), makeErrCallback(customErrHandler), 1)
 }
 
 func (p *producerService) send(topic string, key, value []byte) {
 
 	TopicService.Create(topic)
-	p.create()
+	p.get()
 
 	defer func() {
 		if r := recover(); r != nil {
@@ -34,19 +37,24 @@ func (p *producerService) send(topic string, key, value []byte) {
 			if err := p.producer.Close(); err != nil {
 				tracer.Errorf("dispatcher", "Error closing producer: %v", err.Error())
 			}
+			p.producer = nil
 		}
 	}()
 
+	p.producer.Input() <- &sarama.ProducerMessage{Topic: topic, Key: sarama.ByteEncoder(key), Value: sarama.ByteEncoder(value)}
+
 	select {
-	case p.producer.Input() <- &sarama.ProducerMessage{Topic: topic, Key: sarama.ByteEncoder(key), Value: sarama.ByteEncoder(value)}:
+	// case p.producer.Input() <- &sarama.ProducerMessage{Topic: topic, Key: sarama.ByteEncoder(key), Value: sarama.ByteEncoder(value)}:
+	case msg := <-p.producer.Successes():
+		tracer.Errorf(glob.ProjName, " Sent [%v-%v/%v/%v]\n", msg.Topic, msg.Offset, string(key[:]), glob.TrimBytes(value))
 	case err := <-p.producer.Errors():
 		tracer.Errorf(glob.ProjName, " Failed to produce message: %v", err)
 	}
 
-	tracer.Tracef(glob.ProjName, " Sent: [%v/%v/%v]\n", topic, string(key[:]), string(value[:]))
+	// tracer.Tracef(glob.ProjName, " Sent: [%v/%v/%v]\n", topic, string(key[:]), glob.TrimBytes(value))
 }
 
-func (p *producerService) create() {
+func (p *producerService) get() {
 	if p.producer == nil {
 		p.lock.Lock()
 		if p.producer == nil {
@@ -60,4 +68,23 @@ func (p *producerService) create() {
 		p.lock.Unlock()
 	}
 	return
+}
+
+func makeErrCallback(producerErrHandler model.ProducerCustomerErrHandler) model.ConsumerCallback {
+	defer func() {
+		if err := recover(); err != nil {
+			tracer.Errorf(glob.ProjName, " Panic on err handler: %v", string(debug.Stack()))
+		}
+	}()
+	return func(key, value []byte) (err error) {
+		tracer.Trace(glob.ProjName, "Received err from consumer")
+		var item model.ConsumerCallbackError
+		err = json.Unmarshal(value, &item)
+		if err != nil {
+			tracer.Errorf(glob.ProjName, "Error parsing callbackErr: %v", err.Error())
+			return
+		}
+		producerErrHandler(item.Message.Key, item.Message.Value, errors.New(item.ErrStr))
+		return
+	}
 }